through sub_8124668

+++ include/region_map.h
@@ -309,6 +309,6 @@ u16 GetRegionMapSectionIdAt(u16 x, u16 y);
 void CreateRegionMapPlayerIcon(u16 x, u16 y);
 void CreateRegionMapCursor(u16 tileTag, u16 paletteTag);
 u8 *GetMapName(u8 *, u16, u16);
-bool32 sub_8124668(u16 mapSecId);
+bool32 sub_8124668(u8 mapSecId);
 
 #endif //GUARD_REGION_MAP_H
